+++ libnopegl/src/backends/vk/buffer_vk.h
@@ -35,9 +35,9 @@ struct buffer_vk {
 };
 
 struct buffer *ngli_buffer_vk_create(struct gpu_ctx *gpu_ctx);
-VkResult ngli_buffer_vk_init(struct buffer *s, int size, int usage);
-VkResult ngli_buffer_vk_upload(struct buffer *s, const void *data, int size, int offset);
-VkResult ngli_buffer_vk_map(struct buffer *s, int size, int offset, void **data);
+VkResult ngli_buffer_vk_init(struct buffer *s, size_t size, int usage);
+VkResult ngli_buffer_vk_upload(struct buffer *s, const void *data, size_t size, size_t offset);
+VkResult ngli_buffer_vk_map(struct buffer *s, size_t size, size_t offset, void **data);
 void ngli_buffer_vk_unmap(struct buffer *s);
 void ngli_buffer_vk_freep(struct buffer **sp);
